requirements.txt
python-dotenv
requests
pandas
openpyxl
boto3
botocore